% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_DIF_classification.R
\name{calculate_DIF_classification}
\alias{calculate_DIF_classification}
\title{Calculate Differential Item Functioning (DIF) classification table}
\usage{
calculate_DIF_classification(responses, model = NULL)
}
\arguments{
\item{responses}{A data frame containing the responses to the items.}

\item{model}{Optional. A \code{dmodel} object used to calculate the \code{dscore}.
If not specified, the function will calculate the D-score using the
default key.}
}
\value{
A data frame with items and their classification for DIF.
}
\description{
Classifies each items according to the Jodoin/Gierl criteria for DIF.
Currently implemented:
}
\details{
\itemize{
\item phase (1 vs 2)
\item each country vs others
}
}
