% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calulate_dmodel.R
\name{calculate_dmodel}
\alias{calculate_dmodel}
\title{Calculate the D-score model object given a fitted Rasch model}
\usage{
calculate_dmodel(
  data,
  fit,
  anchors = NULL,
  name = NULL,
  key = NULL,
  transform = c(`(Intercept)` = 0, original = 1),
  itemtable_from = NULL,
  relevance = c(-Inf, Inf),
  population = NULL,
  prior_mean = NULL,
  background_equate = FALSE
)
}
\arguments{
\item{data}{A matrix or a data frame with item responses. The data can take
two shapes. In the \strong{wide shape} the data is a matrix or data frame
with items in columns and visits in the rows. In the \strong{long shape} the data
is a data frame with item responses in long format, i.e. with
visit identifying variables, a column with item names and a column
with responses. The wide format is simpler to work with, but the long
format allows for more flexibility in the data structure and is more
efficient for large datasets. The responses should be coded as \code{1} (pass) or
\code{0} (fail). Missing responses are allowed and are coded as \code{NA}.}

\item{fit}{An object produced by \code{\link{rasch}}.}

\item{anchors}{A named numeric vector of length 2. The names of the vector
must correspond to two item names in \code{data} that will act as anchors
to define the scale units. If specified, the function will update
both \code{tau} and \code{transform} parameters. The default is \code{anchors = NULL}, which
means that the \code{tau} is calculated through the \code{transform} argument.}

\item{name}{If specified, the string will be appended to
"number of items"_"number of equates".}

\item{key}{A string with a provisional key name. If not specified, the
\code{name} is used.}

\item{transform}{A vector of length 2 with the intercept and slope of the
linear transformation of item difficulties \code{tau} to the D-score scale.
The default c(0, 1) makes all calculations in the logit scale.}

\item{itemtable_from}{Location of the \code{itemtable} for the items.
The default \code{NULL} selects \code{dscore::builtin_itemtable}.
Setting \code{itemtable_from} to \code{""} creates simple sequential labels.
Setting \code{itemtable_from} to \code{ddata::itemtable} selects the itemtable
object in the \code{ddata} package. Note that items not listed in \code{itemtable}
will not '' appear in the \code{itembank}.}

\item{relevance}{See \code{\link[dscore]{dscore}}.}

\item{population}{Name of reference to calculate DAZ: See dscore()}

\item{prior_mean}{Name of prior mean column: See dscore()}

\item{background_equate}{A named list or a logical. The named list
has the same structure as the \code{equate} parameter. The argument specifies a
list of all (active and passive) equate groups. Each element specifies
a vector of item names belonging to the same equate group. Setting
\code{background_equate = TRUE} calculates the parameter
from \code{dscore::builtin_itemtable}. The function adds new equate groups
and updates existing equate groups by the corresponding
\code{background_equate}, or overwrite an existing element along
the \code{equate} object for active equates. Setting
\code{background_equate = FALSE} (default) does not specify any passive
equates.}
}
\value{
An object of class \code{dmodel}. See details for a description
of the list elements.
}
\description{
This function calculates ability estimates and diagnostics from
a fitted Rasch model. It prepares the item bank, calculates
the corresponding D-score and DAZ, and computes fit statistics
for items and persons. It also handles equating and anchoring for
the D-score transformation. Results can be returned in either the logit
or in the D-score metric.
}
\examples{
\dontrun{
library("dmetric")
library("tidyr")

# test with Dutch items
varlist <- prepare_items(study = "Netherlands 1")
dmo <- fit_dmodel(varlist, name = "Dutch")
}
}
