% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_palette.R
\name{get_palette}
\alias{get_palette}
\title{Obtains color palettes from a data package}
\usage{
get_palette(
  palet = c("cohort", "study", "country", "domain", "instrument", "wave"),
  na_color = "grey"
)
}
\arguments{
\item{palet}{The name of the list component in the built-in
color palette list \code{gsed_palettes}. Currently implemented
are \code{"cohort"}, \code{"study"}, \code{"cohort"}, \code{"country"},
\code{"domain"}, \code{"wave"} and \code{"instrument"}.}

\item{na_color}{The color that is return if not found. The default is
\code{"grey"}.}
}
\value{
The selected color palette. If not found, it returns the single
color specified by \code{na_color}.
}
\description{
This function returns one of built-in color palettes from a
specified data package.
}
\examples{
\dontrun{
 get_palette("domain")
 }
}
