% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gsed_palettes}
\alias{gsed_palettes}
\title{Color palettes for the gsed data}
\format{
A named \code{list} with vectors.
}
\usage{
gsed_palettes
}
\description{
A list containing six color palettes for the gsed data.
}
\keyword{datasets}
