% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_p_a_item.R
\name{plot_p_a_item}
\alias{plot_p_a_item}
\title{Plot probability by age, all items}
\usage{
plot_p_a_item(pass, item_var = "item", items = NULL, model_name = "", ...)
}
\arguments{
\item{pass}{A \code{data.frame} with emperical probabilities per item per month,
with expected columns \code{item}, \code{cohort}, \code{agegp}, \code{p}, \code{a}, \code{n}, \code{equate}
and \code{label}.}

\item{item_var}{\emph{Only relevant for long shape}. Character string with the
name of the column containing item names. The default is \code{item_var = "item"}.}

\item{items}{Character vector with item names. If not specified,
all columns in the data are included as items (for wide shape) or
all items in the item column are included (for long shape).}

\item{model_name}{A string with the model name, used in the plot title.}

\item{\dots}{Additional parameters passed down to \code{plot_p_a_one_item()}.}
}
\value{
List of \code{ggplot} objects
}
\description{
This function produces a canned \code{ggplot} object.
}
