% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pass.R
\name{plot_pass}
\alias{plot_pass}
\title{Plot passing probability per item, by age or D-score}
\usage{
plot_pass(
  pass,
  data_rug = NULL,
  items = NULL,
  item_var = "item",
  response_var = "response",
  x_var = "a",
  by_var = "cohort",
  min_n = 10,
  model_name = "",
  xlim = NULL,
  xbreaks = NULL,
  xlab = NULL,
  label_trunc = 60,
  col_manual = NULL,
  na_value = "grey"
)
}
\arguments{
\item{pass}{A \code{data.frame} with empirical probabilities per item.
For age, the expected columns are \code{item}, \code{cohort}, \code{p}, \code{a}, \code{n},
\code{equate} and \code{label}.
For D-score, the expected columns are \code{item}, \code{cohort}, \code{dcut}, \code{p}, \code{d},
\code{n}, \code{equate} and \code{label}.}

\item{data_rug}{Optional data frame of individual responses to the item,
used to add rug plots. Should include \code{item}, \code{age}/\code{dscore},
\code{response} and \code{cohort}.}

\item{items}{Character vector with item names. If not specified,
all columns in the data are included as items (for wide shape) or
all items in the item column are included (for long shape).}

\item{item_var}{\emph{Only relevant for long shape}. Character string with the
name of the column containing item names. The default is \code{item_var = "item"}.}

\item{response_var}{\emph{Only relevant for long shape}. Character string with
the name of the column containing responses. The default is
\code{response_var = "response"}.}

\item{x_var}{Variable representing the continuous x-axis
(e.g., age in months, or D-score).}

\item{by_var}{Variable by which to group the data for plotting (e.g.
cohort).}

\item{min_n}{Minimum number of observations required to include a point.
Defaults to 10.}

\item{model_name}{Optional model name used as annotation in the plot.}

\item{xlim}{Numeric vector of length 2 specifying x-axis limits
(in months).}

\item{xbreaks}{Sequence of breaks for the x-axis.}

\item{xlab}{Label for the x-axis.}

\item{label_trunc}{Maximum length of the item label to display.}

\item{col_manual}{Optional named vector of manual colors by \code{cohort}.
If \code{NULL}, colors are obtained using \code{get_palette()}.}

\item{na_value}{Color to use for missing values. Default is \code{"grey"}.}
}
\value{
A \code{ggplot} object.
}
\description{
This function produces a canned \code{ggplot} object.
}
