% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tau_contrast.R
\name{plot_tau_contrast}
\alias{plot_tau_contrast}
\title{Plot tau contrast between two models}
\usage{
plot_tau_contrast(
  model,
  model_old,
  detrended = TRUE,
  ylim = c(-10, 10),
  colors = c("#0072B2", "#D55E00", "#999999"),
  fontcolor = "white"
)
}
\arguments{
\item{model}{A \code{dmodel} object containing the new itembank and item_fit.}

\item{model_old}{A \code{dmodel} object containing the old itembank and item_fit.}

\item{detrended}{Logical, if TRUE, plot the difference in tau values.}

\item{ylim}{Numeric vector of length 2, specifying the y-axis limits.}

\item{colors}{Character vector of colors for the item categories.}

\item{fontcolor}{Character, color for the hover label font.}
}
\value{
A plotly object representing the tau contrast plot.
}
\description{
Plot tau contrast between two models
}
