% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.dmodel}
\alias{print.dmodel}
\title{Print a \code{dmodel} object}
\usage{
\method{print}{dmodel}(x, ...)
}
\arguments{
\item{x}{Object of class \code{dmodel}}

\item{...}{Other parameters passed down to \code{print.default()}}
}
\value{
\code{NULL}
}
\description{
Print a \code{dmodel} object
}
