% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasch.R
\name{rasch}
\alias{rasch}
\title{Pairwise Estimation of the Rasch Model under Item Equating}
\usage{
rasch(
  data,
  shape = c("auto", "wide", "long"),
  visit_var = c("subjid", "agedays"),
  item_var = "item",
  response_var = "response",
  items = NULL,
  equate = NULL,
  b_fixed = NULL,
  b_init = NULL,
  zerosum = FALSE,
  pairs = NULL,
  conv = 1e-05,
  maxiter = 3000,
  progress = FALSE,
  save_pairs = FALSE,
  save_wide = FALSE
)
}
\arguments{
\item{data}{A matrix or a data frame with item responses. The data can take
two shapes. In the \strong{wide shape} the data is a matrix or data frame
with items in columns and visits in the rows. In the \strong{long shape} the data
is a data frame with item responses in long format, i.e. with
visit identifying variables, a column with item names and a column
with responses. The wide format is simpler to work with, but the long
format allows for more flexibility in the data structure and is more
efficient for large datasets. The responses should be coded as \code{1} (pass) or
\code{0} (fail). Missing responses are allowed and are coded as \code{NA}.}

\item{shape}{Character string indicating the shape of the data, either
\code{"wide"}, \code{"long"} or \code{"auto"}. If not specified, the function tries
to determine the shape automatically.}

\item{visit_var}{\emph{Only relevant for long shape}. Character vector with
names of the columns identifying unique visits. Ability is assumed to be
constant for all measurements made during the visit. The default names
are \code{visit_var = c("subjid", "agedays")}.}

\item{item_var}{\emph{Only relevant for long shape}. Character string with the
name of the column containing item names. The default is \code{item_var = "item"}.}

\item{response_var}{\emph{Only relevant for long shape}. Character string with
the name of the column containing responses. The default is
\code{response_var = "response"}.}

\item{items}{Character vector with item names. If not specified,
all columns in the data are included as items (for wide shape) or
all items in the item column are included (for long shape).}

\item{equate}{Optional list with elements. Each list element specifies
a character vector of names of items that should receive identical difficulty
estimates. The default (\code{NULL}) indicates that difficulty estimates
are not restricted to be identical.}

\item{b_fixed}{Numeric, named vector used for fixing the item parameter
estimate to a specific value. The names of the vector indicate
the item name to which the fixed value applies.}

\item{b_init}{Numeric, named vector of initial item difficulty estimates.
Under the default (\code{NULL}) values initial values are calculated
internally.}

\item{zerosum}{Optional logical indicating whether item difficulties
should be centered in each iteration. The default is that no centering
is conducted.}

\item{pairs}{A table of counts \code{t(data == 0)} times \code{data == 1}
for all items specified by \code{items}.
The default (\code{NULL}) calculates the pairs table for the active item
set in the data. This step can take substantial execution time
in large dataset. In that case, it is recommended to pre-compute the
pairs table and pass it to this function.}

\item{conv}{Convergence criterion in maximal absolute parameter change}

\item{maxiter}{Maximal number of iterations}

\item{progress}{A logical which displays the iterative process.
Default is \code{FALSE}.}

\item{save_pairs}{Logical. Save the pairs object in the result?}

\item{save_wide}{Logical. Save the data (wide format) in the result?}
}
\value{
A list with the following elements:
\itemize{
\item \code{item}: A character vector containing the names of the items for which
item parameters were estimated.
\item \code{visit_var}: A character vector with the names of the variables to define
unique visits.
\item \code{shape}: The shape of the data, either \code{"wide"} or \code{"long"}.
\item \code{b_fixed}: The \code{b_fixed} argument values.
\item \code{equate}: The \code{equate} argument values.
\item \code{b_init}: The \code{b_init} argument values.
\item \code{orphans}: A character vector with item names that are not connected
to any other item in the data.
\item \code{zerosum}: The \code{zerosum} argument value.
\item \code{iter}: Iteration counter.
\item \code{convergence}: Convergence criterion.
\item \code{item}: Data frame with estimated item parameters (\code{n}, \code{p}, \code{b}).
\item \code{betapar}: A named vector with negated item difficulties.
\item \code{call}: The matched call to the function.
\item \code{pairs}: The pairs table used by the algorithm, if \code{save_pairs} is \code{TRUE}.
May be used to speed up subsequent calls for the same set of items.
\item \code{wide}: The data in wide format, if \code{save_wide} is \code{TRUE}.
}
}
\description{
This function uses pairwise conditional likelihood estimation for
estimating item parameters in the Rasch model.
}
\details{
This function is loosely based on \code{sirt::rasch.pairwise.itemcluster()}.

The \code{rasch_wide()} and \code{rasch_long()} functions provide a few special
capabilities:
\enumerate{
\item They allow the user to specify a set of items that should be equated.
\item They allow the user to fix item parameters to a specific value.
\item They allow the user to specify a pairs table, which can speed up
subsequent calls for the same set of items.
}
}
\note{
No standard errors are provided by this function. Use resampling methods
for conducting statistical inference. Formulas for asymptotic standard
errors of this pairwise estimation method are described in Zwinderman (1995).
}
\references{
van der Linden, W. J., & Eggen, T. J. H. M. (1986).
\emph{An empirical Bayes approach to item banking}. Research Report 86-6,
University of Twente.

Zwinderman, A. H. (1995). Pairwise parameter estimation in Rasch models.
\emph{Applied Psychological Measurement}, \bold{19}, 369-375.
}
\author{
Stef van Buuren
}
